% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liesel.R
\name{add_copula}
\alias{add_copula}
\title{Add a copula to a pair of marginal distributional regression models}
\usage{
add_copula(
  model0,
  model1,
  copula = "GaussianCopula",
  predictors = list(
    dependence = predictor(~1, inverse_link = "AlgebraicSigmoid")
  ),
  data = NULL,
  builder = FALSE,
  knots = NULL
)
}
\arguments{
\item{model0}{The first marginal distributional regression model.}

\item{model1}{The second marginal distributional regression model.}

\item{copula}{A string identifying a copula in Liesel's TensorFlow
Probability module.}

\item{predictors}{A list of \code{\link[=predictor]{predictor()}} specifications. The names of
the list must match the names of the parameters of the
TensorFlow distribution.}

\item{data}{A data frame or list containing the data for the model.
By default, the data is extracted from the environment of
the formulas.}

\item{builder}{Whether to return the model builder or the model.}

\item{knots}{A list containing the knots per term.
Passed on to \code{\link[mgcv:gam]{mgcv::gam()}}.}
}
\description{
Build a copula regression model from two marginal distributional regression
models from the \code{\link[=liesel]{liesel()}} function.
}
