% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{check_liesel_version}
\alias{check_liesel_version}
\title{Check if the installed versions of RLiesel and Liesel are compatible}
\usage{
check_liesel_version()
}
\description{
This function is defunct. If an incompatible version of Liesel is found,
reticulate will install a compatible version automatically.
See \code{vignette("package", "reticulate")}.
}
