% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_model}
\alias{load_model}
\title{Load a Liesel model from a pickle file}
\usage{
load_model(path)
}
\arguments{
\item{path}{The file path to load the model from.}
}
\description{
Load a Liesel model from a pickle file
}
