% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.liesel.liesel.model.Model}
\alias{plot.liesel.liesel.model.Model}
\title{Plot the graph of a Liesel model}
\usage{
\method{plot}{liesel.liesel.model.Model}(x, nodes = FALSE, ...)
}
\arguments{
\item{x}{The model to plot.}

\item{nodes}{Whether to plot the computational nodes instead of the
statistical variables. Defaults to \code{FALSE}.}

\item{...}{Passed on to the \code{plot_model()} function
in the \code{liesel.liesel.viz} module.}
}
\description{
Plot the graph of a Liesel model
}
