% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liesel.R
\name{predictor}
\alias{predictor}
\title{Set up a Liesel regression predictor specification}
\usage{
predictor(formula = ~1, inverse_link = "Identity", p_smooth_name = NULL)
}
\arguments{
\item{formula}{A one-sided formula.}

\item{inverse_link}{A string identifying a
\href{https://www.tensorflow.org/probability/api_docs/python/tfp/bijectors}{TensorFlow bijector}
to be used as the inverse link.}

\item{p_smooth_name}{The name of the parametric smooth. If \code{NULL},
the name is set automatically.}
}
\description{
Helper function to set up a Liesel regression predictor specification.
}
