% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_model}
\alias{save_model}
\title{Save a Liesel model to a pickle file}
\usage{
save_model(model, path)
}
\arguments{
\item{model}{The model to save.}

\item{path}{The file path to save the model to.}
}
\description{
Save a Liesel model to a pickle file
}
