% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{use_liesel_venv}
\alias{use_liesel_venv}
\title{Use a Liesel virtual environment}
\usage{
use_liesel_venv(virtualenv = NULL, python = NULL, version = NULL)
}
\arguments{
\item{virtualenv}{Either the name of, or the path to, a Python virtual environment.}

\item{python}{The path to a Python binary.}

\item{version}{The version of Python to use. \code{reticulate} will search for versions of
Python as installed by the \code{\link[reticulate:install_python]{install_python()}} helper function.}
}
\description{
This function is defunct. Please configure the Liesel virtual or
conda environment manually, e.g. using \code{\link[reticulate:use_python]{reticulate::use_virtualenv()}}
or \code{\link[reticulate:use_python]{reticulate::use_condaenv()}}. Alternatively, let reticulate manage
the environment automatically. See \code{vignette("package", "reticulate")}.
}
