% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{use_tmp_liesel_venv}
\alias{use_tmp_liesel_venv}
\title{Create and use a temporary Liesel virtual environment}
\usage{
use_tmp_liesel_venv(...)
}
\arguments{
\item{...}{Passed on to \code{\link[reticulate:virtualenv-tools]{reticulate::virtualenv_create()}}.}
}
\description{
Create a temporary Python virtual environment, install Liesel in it,
and use it with \code{reticulate}.
}
\keyword{internal}
