% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{check_liesel_version}
\alias{check_liesel_version}
\title{Check if the installed versions of RLiesel and Liesel are compatible}
\usage{
check_liesel_version()
}
\description{
Note that this function needs to be called after setting the virtual or
conda environment, e.g. using \code{\link[reticulate:use_python]{reticulate::use_virtualenv()}} or
\code{\link[reticulate:use_python]{reticulate::use_condaenv()}}.
}
