% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liesel.R
\name{liesel}
\alias{liesel}
\title{Set up a Liesel distributional regression model}
\usage{
liesel(
  response,
  distribution = "Normal",
  predictors = list(
    loc = predictor(~1, inverse_link = "Identity"),
    scale = predictor(~1, inverse_link = "Exp")
  ),
  data = NULL,
  knots = NULL,
  diagonalize_penalties = TRUE,
  builder = FALSE
)
}
\arguments{
\item{response}{The response vector (or matrix).}

\item{distribution}{A string identifying a
\href{https://www.tensorflow.org/probability/api_docs/python/tfp/distributions}{TensorFlow distribution}
to be used as the response distribution.}

\item{predictors}{A list of \code{\link[=predictor]{predictor()}} specifications. The names of
the list must match the names of the parameters of the
TensorFlow distribution.}

\item{data}{A data frame or list containing the data for the model.
By default, the data is extracted from the environment of
the formulas.}

\item{knots}{A list containing the knots per term.
Passed on to \code{\link[mgcv:gam]{mgcv::gam()}}.}

\item{diagonalize_penalties}{Whether to diagonalize the smooth penalties.}

\item{builder}{Whether to return the model builder or the model.}
}
\description{
Set up a probabilistic graphical model (PGM) representation of a
distributional regression model (also known as generalized additive model
for location, scale, and shape (GAMLSS)) with Liesel.
}
