% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{use_tmp_liesel_venv}
\alias{use_tmp_liesel_venv}
\title{Create and use a temporary Liesel virtual environment}
\usage{
use_tmp_liesel_venv(python = NULL, version = NULL)
}
\arguments{
\item{python}{The path to a Python interpreter, to be used with the created
virtual environment. When \code{NULL}, the Python interpreter associated with
the current session will be used.}
}
\description{
Create a temporary Python virtual environment, install Liesel in it,
and use it with \code{reticulate}.
}
\keyword{internal}
